+++ maml_rl/policies/normal_mlp.py
@@ -36,11 +36,11 @@ class NormalMLPPolicy(Policy):
     def forward(self, input, params=None):
         if params is None:
             params = OrderedDict(self.named_parameters())
-        output = input
+        output = input.float()
         for i in range(1, self.num_layers):
             output = F.linear(output,
-                weight=params['layer{0}.weight'.format(i)],
-                bias=params['layer{0}.bias'.format(i)])
+                weight=params['layer{0}.weight'.format(i)].float(),
+                bias=params['layer{0}.bias'.format(i)].float())
             output = self.nonlinearity(output)
         mu = F.linear(output, weight=params['mu.weight'],
             bias=params['mu.bias'])
